requirements.txt
flask
flask-bcrypt
psycopg2-binary
python-dotenv
